make authorization

+++ app/Http/Controllers/RoleController.php
@@ -6,17 +6,22 @@ use App\Http\Resources\RoleResource;
 use App\Role;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpFoundation\Response;
 
 class RoleController extends Controller
 {
     public function index()
     {
+        Gate::authorize('view','roles');
+
         return RoleResource::collection(Role::all());
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('edit','roles');
+
         $role = Role::create($request->only('name'));
 
          if($permission = $request->input('permissions')) {
@@ -33,11 +38,15 @@ class RoleController extends Controller
 
     public function show($id)
     {
+        Gate::authorize('view','roles');
+
         return new RoleResource(Role::find($id));
     }
 
     public function update(Request $request, $id)
     {
+        Gate::authorize('edit','roles');
+
         $role = Role::find($id);
 
         $role->update($request->only('name'));
@@ -58,6 +67,8 @@ class RoleController extends Controller
 
     public function destroy($id)
     {
+        Gate::authorize('edit','roles');
+
         DB::table('role_permissions')->where('role_id', $id)->delete();
 
         Role::destroy($id);
